Move Gitlab::Satellite into the Satellite module
